Seperate out EigenJacobianTest #REF34278
